MNT #286 utils relocated
