+++ apstools/plans/__init__.py
@@ -1,0 +1,2 @@
+
+from .motors import *
